requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
mysqlclient==2.2.4
PyMySQL==1.1.0
Flask-Migrate==4.0.5
Faker==22.6.0